use cases implemented
