--- app/src/main/java/com/example/marvelapp/data/service/entity/ComicsEntity.kt
@@ -1,5 +1,0 @@
-package com.example.marvelapp.data.service.entity
-
-class ComicsEntity(
-    val available: Int
-)
