--- app/src/main/java/com/example/marvelapp/data/service/entity/EventsEntity.kt
@@ -1,5 +1,0 @@
-package com.example.marvelapp.data.service.entity
-
-class EventsEntity (
-    val available: Int
-)
