--- app/src/main/java/com/example/marvelapp/data/service/entity/SeriesEntity.kt
@@ -1,5 +1,0 @@
-package com.example.marvelapp.data.service.entity
-
-class SeriesEntity (
-    val available: Int
-)
