--- app/src/main/java/com/example/marvelapp/data/service/entity/StoriesEntity.kt
@@ -1,5 +1,0 @@
-package com.example.marvelapp.data.service.entity
-
-class StoriesEntity(
-    val available: Int
-)
