+++ app/src/main/java/com/example/marvelapp/data/serviceImpl/MarvelServiceImpl.kt
@@ -1,9 +1,10 @@
-package com.example.marvelapp.data.service
+package com.example.marvelapp.data.serviceImpl
 
 import com.example.marvelapp.data.request.generator.MarvelRequestGenerator
 import com.example.marvelapp.data.service.api.MarvelApi
-import com.example.marvelapp.data.service.entity.CharacterEntity
-import com.example.marvelapp.data.service.mapper.CharacterMapper.transformListOfCharacters
+import com.example.marvelapp.domain.entity.CharacterEntity
+import com.example.marvelapp.data.mapper.CharacterMapper.transformListOfCharacters
+import com.example.marvelapp.domain.contracts.MarvelService
 import io.reactivex.rxjava3.core.Observable
 
 class MarvelServiceImpl : MarvelService {
