+++ app/src/main/java/com/example/marvelapp/domain/entity/CharacterEntity.kt
@@ -1,4 +1,4 @@
-package com.example.marvelapp.data.service.entity
+package com.example.marvelapp.domain.entity
 
 data class CharacterEntity(
         val id: Int,
