+++ app/src/main/java/com/example/marvelapp/domain/entity/ComicsEntity.kt
@@ -1,0 +1,5 @@
+package com.example.marvelapp.domain.entity
+
+class ComicsEntity(
+    val available: Int
+)
