+++ app/src/main/java/com/example/marvelapp/domain/entity/EventsEntity.kt
@@ -1,0 +1,5 @@
+package com.example.marvelapp.domain.entity
+
+class EventsEntity (
+    val available: Int
+)
