+++ app/src/main/java/com/example/marvelapp/domain/entity/SeriesEntity.kt
@@ -1,0 +1,5 @@
+package com.example.marvelapp.domain.entity
+
+class SeriesEntity (
+    val available: Int
+)
