+++ app/src/main/java/com/example/marvelapp/domain/entity/StoriesEntity.kt
@@ -1,0 +1,5 @@
+package com.example.marvelapp.domain.entity
+
+class StoriesEntity(
+    val available: Int
+)
