+++ app/src/main/java/com/example/marvelapp/domain/usecases/GetCharactersUseCaseImpl.kt
@@ -1,0 +1,13 @@
+package com.example.marvelapp.domain.usecases
+
+import com.example.marvelapp.domain.contracts.MarvelService
+import com.example.marvelapp.domain.entity.CharacterEntity
+import io.reactivex.rxjava3.core.Observable
+
+interface GetCharacterUseCase {
+    fun call(): Observable<List<CharacterEntity>>
+}
+
+class GetCharactersUseCaseImpl constructor(private val marvelService: MarvelService) : GetCharacterUseCase {
+    override fun call(): Observable<List<CharacterEntity>> = marvelService.getCharacters()
+}
